bugfix / new converter for each preview
Instead of creating a single Markdown.Converter object to handle all
windows, this creates a new one for each preview pane. This partially
fixes the behavior when two previews are being used at the same time.

Also changes the @require for MJPDEditing.js to the one in /dev/. The
one in /master/ doesn't exist anymore due to a reversion of an incorrect
merge.