bugfix / run editor after mathjax hook
The markdown editor now does the initial run() after being hooked into
MathJax rendering. This renders any tex that was already in the hooked
textarea.

This fixes the behavior when editing a comment. Now any TeX that was in
the comment gets rendered when the new preview pane is displayed.