bugfix / refresh editor after mathjax hook
The markdown editor now does a refreshPreview() after being hooked into
MathJax rendering. This renders any tex that was already in the hooked
textarea.

This fixes the behavior when editing a comment. Now any TeX that was in
the comment gets rendered when the new preview pane is displayed.

Similar to 7b948edbb06622d3804e444f164e324ce7a8e545